+++ docs/source/varona.platypus.rst
@@ -1,0 +1,7 @@
+varona.platypus
+===========
+
+.. automodule:: varona.platypus
+   :members:
+   :undoc-members:
+   :show-inheritance:
